Restructure, use webpack for distributables
