+++ .gitignore
@@ -1,2 +1,2 @@
-dist
+lib
 node_modules
