+++ src/addons.js
@@ -1,0 +1,8 @@
+import stampit from './';
+import decorator from './utils/decorator';
+
+stampit.addons = {
+  decorator,
+};
+
+export default stampit;
