+++ test/statics.js
@@ -2,7 +2,7 @@ import has from 'lodash/object/has';
 import React from 'react';
 import test from 'tape';
 
-import stampit from '../src/stampit';
+import stampit from '../src';
 
 test('stampit(React, { statics: obj })', (t) => {
   t.plan(1);
